Create findCodeByTile function
