+++ src/chess/helpers/index.js
@@ -5,7 +5,7 @@ export { default as getSide } from './getSide'
 export { default as getFile } from './getFile'
 export { default as findCodeByAxis } from './findCodeByAxis'
 export { default as findCode } from './findCode'
-export { default as findSideByTile } from './findSideByTile'
+export { default as findCodeByTile } from './findCodeByTile'
 export { default as parseTile } from './parseTile'
 export { default as parseCode } from './parseCode'
 export { default as parseSelected } from './parseSelected'
