--- tests/chess/helpers/findSideByTile.spec.js
@@ -1,17 +1,0 @@
-import { findSideByTile } from '~/chess/helpers'
-
-// prettier-ignore
-const snapshot = [
-  'bRa8', 'bNb8', 'bBc8', 'bQd8', 'bKe8', 'bBf8', 'bNg8', 'bRh8',
-  'bPa7', 'bPb7', 'bPc7', 'bPd7', 'bPe7', 'bPf7', 'bPg7', 'bPh7',
-  'wPa3', 'wPb2', 'wPc2', 'wPd2', 'wPe2', 'wPf2', 'wPg2', 'wPh2',
-  'wRa1', 'wNb1', 'wBc1', 'wQd1', 'wKe1', 'wBf1', 'wNg1', 'wRh1'
-]
-
-describe('#findSideByTile', () => {
-  it('find a code by axis', () => {
-    expect(findSideByTile(snapshot)('d2')).toEqual('white')
-    expect(findSideByTile(snapshot)('c8')).toEqual('black')
-    expect(findSideByTile(snapshot)('d5')).toBeUndefined()
-  })
-})
